# docs/swagger.yaml
basePath: /
host: localhost:8000
info:
  contact: {}
  description: orbit transfer
  title: orbits
  version: 0.0-0
paths:
  /orbits:
    get:
      consumes:
      - application/json
      description: Возвращает всех доступные орбиты
      parameters:
      - description: Название орбиты или его часть
        in: query
        name: orbit_name
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            type: ""
      summary: Получение всех орбит со статусом "Доступна"
      tags:
      - orbits
  /orbits/{orbit}:
    get:
      description: Возвращает подробную информацию об орбите по ее названию
      parameters:
      - description: Название орбиты
        in: path
        name: orbit_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Получение детализированной информации об орбите
      tags:
      - orbits
...
schemes:
- http
swagger: "2.0"